# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Todo Camisetas API",
        "description": "API B2B para gestión de camisetas y clientes",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Servidor local Docker"
        }
    ],
    "paths": {
...
        "/api/tallas": {
            "get": {
                "tags": [
                    "Tallas"
                ],
                "summary": "Listar todas las tallas",
                "operationId": "2fc6784708cffdcf1982d8320f94f8e8",
                "responses": {
                    "200": {
                        "description": "Lista de tallas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nombre": {
                                                "type": "string",
                                                "example": "M"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tallas"
                ],
                "summary": "Crear una talla",
                "operationId": "2e756c6383b2fe39e7e3fa2e72f2465f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "XXXL"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Talla creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 7
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "XXXL"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Nombre de talla duplicado o inválido"
                    }
                }
            }
        },
        "/api/tallas/{id}": {
            "get": {
                "tags": [
                    "Tallas"
                ],
                "summary": "Ver una talla por ID",
                "operationId": "05b4ea823c0b92818f2f4018160a39c2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Talla encontrada con camisetas asociadas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "S"
                                        },
                                        "camisetas": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "titulo": {
                                                        "type": "string",
                                                        "example": "Camiseta Local 2025"
                                                    },
                                                    "codigo_producto": {
                                                        "type": "string",
                                                        "example": "SCL2025L"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Talla no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tallas"
                ],
                "summary": "Actualizar una talla",
                "operationId": "a4a10794a4c3363f4cbb8322e3b0fda3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "XXXL"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Talla actualizada"
                    },
                    "404": {
                        "description": "Talla no encontrada"
                    },
                    "422": {
                        "description": "Nombre de talla duplicado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tallas"
                ],
                "summary": "Eliminar una talla",
                "operationId": "0a6fbc467adf9ed1d3e10aa97639db12",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Talla eliminada"
                    },
                    "404": {
                        "description": "Talla no encontrada"
                    },
                    "422": {
                        "description": "La talla tiene camisetas asociadas y no puede eliminarse"
                    }
                }
            }
        }
    },
...
}